refactor(auth): move auth action into  folder action/auth
