+++ components/auth/form/reset-form.tsx
@@ -1,85 +1,85 @@
-'use client';
-import { zodResolver } from '@hookform/resolvers/zod';
-import { useState, useTransition } from 'react';
-import { useForm } from 'react-hook-form';
-import * as z from 'zod';
+"use client";
+import { zodResolver } from "@hookform/resolvers/zod";
+import { useState, useTransition } from "react";
+import { useForm } from "react-hook-form";
+import * as z from "zod";
 
-import { Button } from '@/components/ui/button';
+import { Button } from "@/components/ui/button";
 import {
-	Form,
-	FormControl,
-	FormField,
-	FormItem,
-	FormLabel,
-	FormMessage,
-} from '@/components/ui/form';
-import { Input } from '@/components/ui/input';
+  Form,
+  FormControl,
+  FormField,
+  FormItem,
+  FormLabel,
+  FormMessage,
+} from "@/components/ui/form";
+import { Input } from "@/components/ui/input";
 
-import { reset } from '@/lib/actions/reset';
-import { CardWraper } from '@/components/auth/ui/card-wraper';
-import { FormError } from '@/components/auth/ui/form-error';
-import { FormSuccess } from '@/components/auth/ui/form-success';
-import { ResetSchema } from '@/lib/schemas';
+import { CardWraper } from "@/components/auth/ui/card-wraper";
+import { FormError } from "@/components/auth/ui/form-error";
+import { FormSuccess } from "@/components/auth/ui/form-success";
+import { reset } from "@/lib/actions/auth/reset";
+import { ResetSchema } from "@/lib/schemas";
 
 export const ResetForm = () => {
-	const [error, setError] = useState<string | undefined>('');
-	const [success, setSuccess] = useState<string | undefined>('');
-	const [isPending, startTransition] = useTransition();
+  const [error, setError] = useState<string | undefined>("");
+  const [success, setSuccess] = useState<string | undefined>("");
+  const [isPending, startTransition] = useTransition();
 
-	const form = useForm<z.infer<typeof ResetSchema>>({
-		resolver: zodResolver(ResetSchema),
-		defaultValues: {
-			email: '',
-		},
-	});
+  const form = useForm<z.infer<typeof ResetSchema>>({
+    resolver: zodResolver(ResetSchema),
+    defaultValues: {
+      email: "",
+    },
+  });
 
-	const onSubmit = (values: z.infer<typeof ResetSchema>) => {
-		setError('');
-		setSuccess('');
+  const onSubmit = (values: z.infer<typeof ResetSchema>) => {
+    setError("");
+    setSuccess("");
 
-		startTransition(() => {
-			reset(values).then((data) => {
-				setError(data?.error);
-				setSuccess(data?.success);
-			});
-		});
-	};
+    startTransition(() => {
+      reset(values).then((data) => {
+        setError(data?.error);
+        setSuccess(data?.success);
+      });
+    });
+  };
 
-	return (
-		<CardWraper
-			headerLabel="Forgot your password?"
-			backButtonLabel="Back to login"
-			backButtonHref="/auth/login"
-		>
-			<Form {...form}>
-				<form onSubmit={form.handleSubmit(onSubmit)} className="space-y-6">
-					<div className="space-y-4">
-						<FormField
-							control={form.control}
-							name="email"
-							render={({ field }) => (
-								<FormItem>
-									<FormLabel>Email</FormLabel>
-									<FormControl>
-										<Input
-											{...field}
-											placeholder="john.doe@example.com"
-											type="email"
-											disabled={isPending}
-										/>
-									</FormControl>
-									<FormMessage />
-								</FormItem>
-							)}
-						/>
-					</div>
-					<FormSuccess message={success} />
-					<FormError message={error} />
-					<Button type="submit" className="w-full" disabled={isPending}>
-						Send reset email
-					</Button>
-				</form>
-			</Form>
-		</CardWraper>
-	);
+  return (
+    <CardWraper
+      headerLabel="Forgot your password?"
+      backButtonLabel="Back to login"
+      backButtonHref="/auth/login"
+    >
+      <Form {...form}>
+        <form onSubmit={form.handleSubmit(onSubmit)} className="space-y-6">
+          <div className="space-y-4">
+            <FormField
+              control={form.control}
+              name="email"
+              render={({ field }) => (
+                <FormItem>
+                  <FormLabel>Email</FormLabel>
+                  <FormControl>
+                    <Input
+                      {...field}
+                      placeholder="john.doe@example.com"
+                      type="email"
+                      disabled={isPending}
+                    />
+                  </FormControl>
+                  <FormMessage />
+                </FormItem>
+              )}
+            />
+          </div>
+          <FormSuccess message={success} />
+          <FormError message={error} />
+          <Button type="submit" className="w-full" disabled={isPending}>
+            Send reset email
+          </Button>
+        </form>
+      </Form>
+    </CardWraper>
+  );
 };
